// backend/src/common/enums/notification.enum.ts
export enum NotificationType {
  BLOCK_USER = 'block user',
  USER_VERIFICATION = 'user verification',
  USER_VERIFICATION_APPROVED = 'user verification approved',
  USER_VERIFICATION_REJECTED = 'user verification rejected',
  SERVICE_REVIEW = 'service review',
  SERVICE_APPROVED = 'service approved',
  SERVICE_REJECTED = 'service rejected',
  CONTRACT_MESSAGE = 'contract message',
  CONTRACT_EXPIRED = 'contract expired',
  CONTRACT_COMPLETED = 'contract completed',
  CONTRACT_DELIVERED = 'contract delivered',
  REVIEW_CREATED = 'review created',
  CONTRACT = 'contract',
  PAYMENT = 'payment',
}